[LLVM] Make -use-constant-fp-for-scalable-splat the default.
Includes a "fix" to ReassociatePass::OptimizeAdd to force the type
of new ConstantFP nodes. A say "fix" because I have not been able to
exercise the affected code in a way that shows the expected error.
However, the fix is trivial and effectively NFC for the original
scalar only use case.